change to integer except decimal

+++ database/migrations/2023_12_07_130010_create_cars_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('model');
             $table->string('kind');
             $table->bigInteger('price');
-            $table->decimal('lowest_down_payment');
+            $table->integer('lowest_down_payment');
             $table->unsignedBigInteger('brand_id');
             $table->foreign('brand_id')->references('id')->on('brands')->onDelete('cascade');
             $table->unsignedBigInteger('user_id');
